+++ tests/integration_tests/functional/test_metrics.py
@@ -33,6 +33,7 @@ def test_flush_metrics(test_microvm_with_api):
         'block',
         'get_api_requests',
         'i8042',
+        'latencies_us',
         'logger',
         'mmds',
         'net',